QA: organize HU reservation code in Repo, Services etc
https://github.com/metasfresh/metasfresh/issues/5035
