+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/IHandlingUnitsDAO.java
@@ -383,4 +383,6 @@ public interface IHandlingUnitsDAO extends ISingletonService
 	List<I_M_Warehouse> retrieveWarehousesWhichContainNoneOf(List<I_M_HU> hus);
 
 	List<I_M_HU> retrieveByIds(Collection<HuId> huIds);
+
+	void setReservedByHUIds(final Collection<HuId> huIds, boolean reserved);
 }
